renamed custom_metrics to metrics
